# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Link Crew Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a school (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: Created school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{id}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get a school by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

  # Auth
...
  /users/me:
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Quiz
...
  /quiz/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz answers and generate interest profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result and interest profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        state:
          type: string
        logoUrl:
          type: ["string", "null"]
        colors:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, city, state, memberCount, createdAt]
    SchoolInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        state:
          type: string
        logoUrl:
          type: ["string", "null"]
        colors:
          type: ["string", "null"]
      required: [name, city, state]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [freshman, leader, admin]
        schoolId:
          type: integer
        schoolName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        privacyLevel:
          type: string
          enum: [public, school_only, private]
        quizCompleted:
          type: boolean
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, schoolId, schoolName, privacyLevel, quizCompleted, createdAt]
...
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        privacyLevel:
          type: string
          enum: [public, school_only, private]
...
    QuizAnswer:
      type: object
      properties:
        questionId:
          type: integer
        answer:
          type: string
      required: [questionId, answer]
    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required: [answers]
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        hobbies:
          type: array
          items:
            type: string
        academicInterests:
          type: array
          items:
            type: string
        extracurriculars:
          type: array
          items:
            type: string
        collegeGoals:
          type: array
          items:
            type: string
        careerInterests:
          type: array
          items:
            type: string
        isExploring:
          type: boolean
        completedAt:
          type: string
      required: [id, userId, hobbies, academicInterests, extracurriculars, collegeGoals, careerInterests, isExploring, completedAt]
...