# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Link Crew Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdatePrivacyLevel } from './userUpdatePrivacyLevel';

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  privacyLevel?: UserUpdatePrivacyLevel;
}